requirements.txt
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
websocket-client==1.7.0
python-dotenv==1.0.0
ccxt==4.2.25
ta-lib==0.4.28
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
krakenex==2.2.2
pykrakenapi==0.3.2
python-dateutil==2.8.2
pytz==2023.3
flask==3.0.0
flask-cors==4.0.0
